+++ Content.Shared/Weapons/Ranged/Components/GunComponent.cs
@@ -263,11 +263,6 @@ public sealed partial class GunComponent : Component
     /// </summary>
     [DataField]
     public Vector2 DefaultDirection = new Vector2(0, -1);
-
-    // RAYTEN
-    // Отвечает за попап если магазин пустой
-    [DataField]
-    public bool PopupOnEmpty = true;
 }
 
 [Flags]
