+++ Resources/Prototypes/Vanilla/Entities/Objects/Weapons/Guns/Battery/portalgun.yml
@@ -34,7 +34,6 @@
   - type: Gun
     soundEmpty: null
     showExamineText: false
-    popupOnEmpty: false
   - type: SolutionContainerManager
     solutions:
       portal:
